rag-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
elasticsearch[async]==8.13.2
anthropic==0.40.0
pydantic-settings==2.4.0
requests==2.32.3
beautifulsoup4==4.12.3
